Rework index search to reduce the number of calls into re2.
re2 has some fixed overhead associated with each call, and we can get
better performance out of feeding it large pieces of text at a
time. If we're searching for something that is relatively dense in the
index, we can get a noticeable boost by passing multiple consecutive
lines at once to RE2, instead of feeding it a line at a time.

We also get to avoid many of the find_line calls, which are pretty
cheap but can add up.