Indexer: Improve support for repetition constructs.
Add a special anchor flag that indicates that the anchor is
repetition. This flag is only meaningful if kAnchorBoth is set, in
which case it indicates that the index can anchor at *either* the
left-hand-side *or* the right-hand-side, but that commiting to one
anchor (specifically, in Concat) destroys its ability to anchor at the
other side.

This flag can be provided to repetition constructs (i.e. + or {N}),
increasing the chance that we'll be able to attach them to an adjacent
index term.