Remove the `lengths' cache from radix_sorter.
This is unnecessary:- For indexing, we don't need it, since the radix
sort automatically stops sorting once it runs off the end of a a
string, and for comparison, we are doing a O(n) scan anyways, so the
data is in cache and adding a memchr() (empirically) does not have a
measurable slowdown.

And since we did one such allocation per thread, this dramatically
improves memory usage when indexing, especially as we bump the chunk
size.