Allow HTML template customization via server configuration.
This commit lets the user inject HTML into the
layout.html template to customize the page.  The
intent is to be able to add additional links or
to otherwise monkey-patch the UI in a lightweight
way via the config file.

The template engine (http://golang.org/pkg/html/template/) is
clever and parses the HTML to escape its inputs, and assumes
the inputs are untrusted.  In this case, it's fair to say
that the server configuration is trusted (and intended
to be passed through), which is why the type in config.go
is "template.HTML".