Handle "other" githubs; fix github links.
With this commit, a user specifying a valid
URL for their github index metadata will
link directly to it, rather than to github.com.
This is useful for organizations that
run Github Enterprise / Github FI on-premise.

As far as I could tell, Github linking only
worked for the first backend, because the
server never sent back the backend as part
of a query response.  I've fixed this
quickly, by looking up the backend in the
in_flight query.  I'm not certain that this
is correct.

I've also extended shorten(ref) to shorten
"origin/master" to "master."  It's hard to
make all possible git references work when
plugged into the github URL namespace, but
this one is common enough that it makes sense.
(And if you named a branch origin/master,
you're already in plenty of pain.)  I could
have made it more explicit by requiring the
configuration to specify refs/remotes/origin/master,
but that seemed too irksome.