Add livegrep-fetch-reindex helper command
Inspired by `livegrep-github-reindex`, this clones/fetches a number of
git repositories in parallel and then regenerates the index. Repository
information is provided in the index json config format used by
`codesearch`, with a `remote` field in the `metadata` dictionary
indicating where to fetch the repository from.

The code is almost entirely plagiarized from `livegrep-github-reindex`.
`livegrep-github-index` itself could probably be refactored to generate
an index config with `remote` fields and then invoke this.