Add repository browser
This adds the ability to view source files directly in livegrep, as
an alternative to linking to external viewers configured by
`url-pattern`. This is desirable for a couple of reasons:

 - It's a common pattern to search for something in the codebase,
   find it, examine its surroundings, and search again for something
   nearby (for example, when following a chain of function calls).

 - The external viewer might be slow (for example, if it's Phabricator),
   and it's valuable to get a snappy response when clicking through to
   view a search result.

The file browser is implemented as part of `livegrep`, and requires
having git repositories available to the `livegrep` process. Therefore,
`livegrep` now accepts an additional config file (in the same format
used by `codesearch`) specifying where to find repos on disk. If this
config is not provided, we retain the old behaviour where clicking
through to view a search result goes directly to the external site.