Use css content: to uncopiably show line numbers
With this approach, line numbers are not included when code is selected
and copied, in at least Chrome and Firefox.

I used the `aria-label` attribute to hold the line number so that the
link is described in a way accessible to screen readers, though I'm not
sure this is the best solution.