Minor css change to allow users to highlight the file path more easily.
Currently, the <a> tag takes up the entire length of the header div.
This means that you can't highlight the file path if you click down
and drag left to highlight the path from the right.

This change makes the <a> tag take up only the minimum width needed,
allowing highlighting to work as expected. The background color
change is done to merge the two backgrounds and produce the same
color as before.