From file view, "enter" opens search in new tab
This is meant to smooth the flow where a user wants to trace multiple
layers of a call stack, and then switch back and forth between
inspecting the function bodies.

The change will be a bit disruptive for users accustomed to "enter"
performing a search in the current tab; hopefully it shouldn't be too
hard for them to switch to "/".