Add buttons to narrow results by file extensions
A popular request from users: buttons that automatically narrow the
current search to files matching a single extension.  The extensions
offered are all culled from the currently displayed files.

Because these buttons need to advance the search history URL to a new
state from which the user often wants to return by hitting Back, this
diff also fixes livegrep's previous problem of not preserving browser
URL history; we now create a browser history state for each search
that successfully returns and is drawn on the screen.