Fix Markdown code snippets piled at top of page
A very generic CSS class name (“code”) was used in the file viewer to
position the pane of source code absolutely at the top of the page.
Alas, “code” is also used as a CSS class when prism.js is highlighting a
code snippet inside of Markdown.  So let’s fix the collision.