Move core logic of no-index to fetch-reindex (#322)
* Move core logic of no-index to fetch-reindex

The original PR implemted the logic in github-reindex not realizing that the fetching of the repos was tightly
coupled to fetch-reindex. This means that, with the old implementation, you could run the config generation but nothing
would actually update. You would just have HEAD but the repos would still be outdated.

Ideally we could de-couple github operations from indexing operations and independently fetch/index, but saving a full refactor
the easiest solution is to pass the flag into fetch-reindex and exit after fetching repos

* Format bool as string

* conditionally append --no-index flag

Co-authored-by: Alex Suraci <vito@stripe.com>